eblan22814882281488

+++ database/migrations/2021_02_26_094144_create_users_table.php
@@ -19,7 +19,7 @@ class CreateUsersTable extends Migration
             $table->string('email')->unique();
             $table->string('password');
             $table->timestamps();
-            $table->unsignedInteger('phone_number')->unique();
+            $table->unsignedBigInteger('phone_number')->unique();
             $table->boolean('notifications')->default(0);
             $table->boolean('auto_update')->default(0);
             $table->boolean('auto_pay')->default(0);
